requirements.txt
Pillow
opencv-python
numpy
matplotlib
scipy
line-profiler
keras
tensorflow
